Move ProfileWizard component out of Dashboard.
