+++ client/profile-wizard/steps/benefits/images/management.js
