+++ client/profile-wizard/steps/benefits/logo.js
