+++ client/profile-wizard/steps/theme/preview.js
