+++ client/profile-wizard/steps/theme/uploader.js
